Shader border size more precision

+++ packages/internal/draw.go
@@ -102,7 +102,7 @@ func QueueTexture(tex rl.Texture2D, src, dst rl.Rectangle, ang float32, col rl.C
 		polygonBuf[i].V2 = packV2(eff.BorderColor)
 		polygonBuf[i].NX = packNormalX(eff.Gamma, eff.Saturation, eff.Contrast, eff.Brightness)
 		polygonBuf[i].NY = packNormalY(0.05, number.Limit(eff.PixelSize, 0, 16), eff.BlurX, eff.BlurY)
-		polygonBuf[i].NZ = packNormalZ(eff.DepthZ, int16(eff.BorderSize), 1)
+		polygonBuf[i].NZ = packNormalZ(eff.DepthZ, eff.BorderSize, 1)
 
 		if true { // sprite
 			polygonBuf[i].TX = packTangentXSprite(eff.OutlineColor)
